refactor: move the resonance dialog into applications
共鳴判定のダイアログを applications/dialogs/ に分離し、インラインの
HTML文字列を templates/apps/resonance-roll.hbs に移した。
文言はこのコミットでは変えていない（次のコミットでローカライズする）。

DialogV2.prompt への巨大なインライン型アサーションも整理した。
prompt の戻り値は本体JSDocで Promise<any> なので受け側で型を付ければよく、
キャストが要るのは config.ok が引数の型に含まれていない点だけだった。

入力値の検証（NaN・0以下を1に丸める）はダイアログのcallback内にあり、
プログラムから intensity を渡した場合には効いていなかった。
rules/resonance-roll.ts の normalizeIntensity に移してある。

EmokloreActor#rollResonance は入力を集めて判定を流すだけになり、
actor.ts 全体で172行→105行。
