+++ templates/apps/resonance-roll.hbs
@@ -1,0 +1,9 @@
+<div>
+  <label for="intensity">強度</label>
+  <input name="intensity" id="intensity" type="number" placeholder="1" min="1" max="9" autofocus>
+</div>
+<div>
+  <label><input type="radio" name="choice" value="none" checked> 一致なし</label>
+  <label><input type="radio" name="choice" value="root"> ルーツ属性一致</label>
+  <label><input type="radio" name="choice" value="completely"> 完全一致</label>
+</div>
